--- src/main/java/com/programmers/springbootbasic/config/properties/FileProperties.java
@@ -1,12 +1,0 @@
-package com.programmers.springbootbasic.config.properties;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@Getter
-@RequiredArgsConstructor
-@ConfigurationProperties(prefix = "app.file")
-public class FileProperties {
-    private final String customers;
-}
