--- src/main/java/com/programmers/springbootbasic/domain/voucher/Voucher.java
@@ -1,8 +1,0 @@
-package com.programmers.springbootbasic.domain.voucher;
-
-import java.util.UUID;
-
-public interface Voucher {
-    UUID getId();
-    long getAmount();
-}
