--- src/main/java/com/programmers/springbootbasic/infra/ApplicationListenerImpl.java
@@ -1,19 +1,0 @@
-package com.programmers.springbootbasic.infra;
-
-import org.springframework.boot.context.event.ApplicationStartedEvent;
-import org.springframework.context.ApplicationListener;
-import org.springframework.stereotype.Component;
-
-@Component
-public class ApplicationListenerImpl implements ApplicationListener<ApplicationStartedEvent> {
-    @Override
-    public void onApplicationEvent(ApplicationStartedEvent event) {
-        System.out.println();
-        System.out.println("=== Voucher Program ===");
-        System.out.println("Type [exit] to exit the program.");
-        System.out.println("Type [create] to create a new voucher.");
-        System.out.println("Type [list] to list all vouchers.");
-        System.out.println("=== Customer Program ===");
-        System.out.println("Type [blacklist] to list all blacklisted customers.");
-    }
-}
