--- src/main/java/com/programmers/springbootbasic/util/IdProvider.java
@@ -1,5 +1,0 @@
-package com.programmers.springbootbasic.util;
-
-public interface IdProvider<T> {
-    T generateId();
-}
