+++ src/main/java/com/programmers/vouchermanagement/VoucherManagementApp.java
@@ -1,0 +1,14 @@
+package com.programmers.vouchermanagement;
+
+import com.programmers.vouchermanagement.config.properties.FileProperties;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+
+@SpringBootApplication
+@EnableConfigurationProperties(FileProperties.class)
+public class VoucherManagementApp {
+    public static void main(String[] args) {
+        SpringApplication.run(VoucherManagementApp.class, args);
+    }
+}
