+++ src/main/java/com/programmers/vouchermanagement/config/properties/FileProperties.java
@@ -1,0 +1,12 @@
+package com.programmers.vouchermanagement.config.properties;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@Getter
+@RequiredArgsConstructor
+@ConfigurationProperties(prefix = "app.file")
+public class FileProperties {
+    private final String customers;
+}
