+++ src/main/java/com/programmers/vouchermanagement/domain/voucher/FixedAmountVoucher.java
@@ -1,0 +1,21 @@
+package com.programmers.vouchermanagement.domain.voucher;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+import java.util.UUID;
+
+@Getter
+@RequiredArgsConstructor
+public class FixedAmountVoucher implements Voucher {
+    private final UUID id;
+    private final long amount;
+
+    @Override
+    public String toString() {
+        return "FixedAmountVoucher{" +
+                "id=" + id +
+                ", amount=" + amount +
+                '}';
+    }
+}
