+++ src/main/java/com/programmers/vouchermanagement/domain/voucher/PercentDiscountVoucher.java
@@ -1,0 +1,21 @@
+package com.programmers.vouchermanagement.domain.voucher;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+import java.util.UUID;
+
+@Getter
+@RequiredArgsConstructor
+public class PercentDiscountVoucher implements Voucher {
+    private final UUID id;
+    private final long amount;
+
+    @Override
+    public String toString() {
+        return "PercentDiscountVoucher{" +
+                "id=" + id +
+                ", amount=" + amount +
+                '}';
+    }
+}
