+++ src/main/java/com/programmers/vouchermanagement/domain/voucher/Voucher.java
@@ -1,0 +1,8 @@
+package com.programmers.vouchermanagement.domain.voucher;
+
+import java.util.UUID;
+
+public interface Voucher {
+    UUID getId();
+    long getAmount();
+}
