+++ src/main/java/com/programmers/vouchermanagement/repository/customer/CustomerRepository.java
@@ -1,0 +1,9 @@
+package com.programmers.vouchermanagement.repository.customer;
+
+import com.programmers.vouchermanagement.domain.customer.Customer;
+
+import java.util.List;
+
+public interface CustomerRepository {
+    List<Customer> findAllBlacklisted();
+}
