+++ src/main/java/com/programmers/vouchermanagement/repository/voucher/MemoryVoucherRepository.java
@@ -1,0 +1,21 @@
+package com.programmers.vouchermanagement.repository.voucher;
+
+import com.programmers.vouchermanagement.domain.voucher.Voucher;
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Repository;
+
+import java.util.*;
+
+@Repository
+@Profile("local")
+public class MemoryVoucherRepository implements VoucherRepository {
+    private final Map<UUID, Voucher> storage = new HashMap<>();
+
+    public void save(Voucher voucher) {
+        storage.put(voucher.getId(), voucher);
+    }
+
+    public List<Voucher> findAll() {
+        return new ArrayList<>(storage.values());
+    }
+}
