+++ src/main/java/com/programmers/vouchermanagement/service/CustomerService.java
@@ -1,0 +1,18 @@
+package com.programmers.vouchermanagement.service;
+
+import com.programmers.vouchermanagement.domain.customer.Customer;
+import com.programmers.vouchermanagement.repository.customer.CustomerRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class CustomerService {
+    private final CustomerRepository customerRepository;
+
+    public List<Customer> blacklist() {
+        return customerRepository.findAllBlacklisted();
+    }
+}
