+++ src/main/java/com/programmers/vouchermanagement/util/IdProvider.java
@@ -1,0 +1,5 @@
+package com.programmers.vouchermanagement.util;
+
+public interface IdProvider<T> {
+    T generateId();
+}
