+++ src/main/java/com/programmers/vouchermanagement/util/UuidProvider.java
@@ -1,0 +1,13 @@
+package com.programmers.vouchermanagement.util;
+
+import org.springframework.stereotype.Component;
+
+import java.util.UUID;
+
+@Component
+public class UuidProvider implements IdProvider<UUID> {
+    @Override
+    public UUID generateId() {
+        return UUID.randomUUID();
+    }
+}
