--- src/test/java/com/programmers/springbootbasic/SpringbootBasicApplicationTests.java
@@ -1,13 +1,0 @@
-package com.programmers.springbootbasic;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringbootBasicApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
