+++ src/test/java/com/programmers/vouchermanagement/VoucherManagementAppTests.java
@@ -1,0 +1,13 @@
+package com.programmers.vouchermanagement;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class VoucherManagementAppTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
