+++ src/test/java/com/programmers/vouchermanagement/mock/util/MockUuidProvider.java
@@ -1,0 +1,16 @@
+package com.programmers.vouchermanagement.mock.util;
+
+import com.programmers.vouchermanagement.util.IdProvider;
+
+import java.util.UUID;
+
+public class MockUuidProvider implements IdProvider<UUID> {
+    private final UUID fixture;
+
+    public MockUuidProvider(UUID fixture) {
+        this.fixture = fixture;
+    }
+    public UUID generateId() {
+        return fixture;
+    }
+}
